Move label formatter to virsat graphiti project - (Task #891)
