+++ de.dlr.sc.virsat.model.extension.statemachines.ui/META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: VirSat Core Model Extension State Machines UI
 Bundle-SymbolicName: de.dlr.sc.virsat.model.extension.statemachines.ui;singleton:=true
-Bundle-Version: 4.12.2.qualifier
+Bundle-Version: 4.12.1.qualifier
 Require-Bundle: de.dlr.sc.virsat.model.extension.statemachines,
  org.eclipse.core.runtime,
  org.eclipse.core.expressions,
@@ -16,8 +16,7 @@ Require-Bundle: de.dlr.sc.virsat.model.extension.statemachines,
  de.dlr.sc.virsat.graphiti.ui,
  de.dlr.sc.virsat.graphiti,
  org.eclipse.nebula.widgets.tablecombo,
- de.dlr.sc.virsat.model.extension.ps,
- org.apache.commons.lang
+ de.dlr.sc.virsat.model.extension.ps
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Bundle-Activator: de.dlr.sc.virsat.model.extension.statemachines.ui.Activator
